% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ELSD.R
\name{calc_ELSD}
\alias{calc_ELSD}
\title{Calculate Earliest to Latest Segment Delay}
\usage{
calc_ELSD(data, subset.pattern = NULL)
}
\arguments{
\item{data}{Dataframe with one column for each segment.}

\item{subset.pattern}{Optional, if e.g dataframe holds data for different slices, eg. basal or midventricular.}
}
\value{
Numeric
}
\description{
Given a dataframe with columns for each segment, each containing time
to peak contraction, this function converts the data to long format and calculates
the difference in time between peak contraction in the earliest and latest segment.
}
\examples{

}
