% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureStrain.R
\name{cureStrain}
\alias{cureStrain}
\title{Circumferential uniformity ratio estimate (CURE) averaged over all timeframes.}
\usage{
cureStrain(data, id.column)
}
\arguments{
\item{data}{Dataframe with id column, and columns for each strain
segment. Each column should represent a separate segment. The segments should
be in the following order: Antero-septal, Infero-septal, Inferior, Infero-lateral,
Antero-lateral and Anterior. Or equivalent.}

\item{id.column}{name of the id column in quotes.}
}
\value{
Returns a dataframe with average CURE value for each patient.
}
\description{
Circumferential uniformity ratio estimate (CURE) averaged over all timeframes.
}
\examples{
cureStrain(data = strain_data, id.column = "id")

}
