% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cureSvd.R
\name{cureSvd}
\alias{cureSvd}
\title{Circumferential uniformity ratio estimate (CURE) computed by singular value decomposition (SVD).}
\usage{
cureSvd(data, id.column)
}
\arguments{
\item{data}{Dataframe with id column, and columns for each strain
segment. Each column should represent a separate segment. The segments should
be in the following order: Antero-septal, Infero-septal, Inferior, Infero-lateral,
Antero-lateral and Anterior. Or equivalent.}

\item{id.column}{name of the id column in quotes.}
}
\value{
Returns a dataframe with CURE value for each patient.
}
\description{
Circumferential uniformity ratio estimate (CURE) computed by singular value decomposition (SVD).
}
\examples{

}
