% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStrainPeaks.R
\name{findStrainPeaks}
\alias{findStrainPeaks}
\title{Find strain peaks using percentage criteria for circumferential/longitudinal strain}
\usage{
findStrainPeaks(x, strain, thresh, incr)
}
\arguments{
\item{x}{dataframe containing the strain columns in a long format.}

\item{strain}{the strain variable you want to mutate, written inside "".}

\item{thresh}{the minimum strain threshold value a valley has to reach to be
considered.}

\item{incr}{by how much must the strain increase between valley and the
following peak.}
}
\value{
A dataframe with id and a straincolumn containing the time for the first peak
fulfilling the criteria
}
\description{
Find strain peaks using percentage criteria for circumferential/longitudinal strain
}
\examples{
findStrainPeaks(strain_data, "mas", thresh = 0.2, incr = -0.3)
}
