% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_cure.R
\name{strain_cure_svd}
\alias{strain_cure_svd}
\title{Circumferential uniformity ratio estimate from singular value
decomposition}
\usage{
strain_cure_svd(data)
}
\arguments{
\item{data}{Dataframe, columns containing strainvalues from LV wall segments
according to the AHA-17 segment model.}
}
\description{
Circumferential uniformity ratio estimate based on singular value decomposition
of cardiac strain for dyssynchrony analysis.
}
\section{Background}{

Circumferential uniformity ratio estimate (cure) is a measure typically used
for analysis of left ventricular (LV) mechanical dyssynchrony. Ranging from 0,
meaning complete dyssynchrony, that is, all opposing wall segments in LV move
in different directions. And 1 meaning complete synchrony, all walls contract
and shortens at the same time. Cure \eqn{\le 0.75} has previously been used to
define dyssynchrony.
}

