% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_elsd.R
\name{strain_elsd}
\alias{strain_elsd}
\title{Earliest-to-latest segment delay}
\usage{
strain_elsd(data, subset_pattern = NULL)
}
\arguments{
\item{data}{Dataframe, columns containing strainvalues from LV wall segments
according to the AHA-17 segment model.}

\item{subset_pattern}{Regex pattern for subsetting over which columns to perform
analysis.}
}
\description{
Time delay between earliest and latest contracting left ventricular wall
segments
}
\details{
Non-uniform contraction of the left ventricular wall can be quantified by
means of dyssynchrony analysis. It is known that in case of left bundle branch
block, the septum normally contracts before the lateral free wall. However,
this is not always the case and therefore measureing the delay between earliest
and latest segment, independent of location might better characterize the
pattern of dyssynchrony.
}
