% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeMinStrain.R
\name{timeMinStrain}
\alias{timeMinStrain}
\title{Find the time corresponding to the minimum strain value for a segment}
\usage{
timeMinStrain(x, segment)
}
\arguments{
\item{x}{Dataframe with strain data in a long format, one column for each
segment.}

\item{segment}{Name of the strain segment.}
}
\value{
A dataframe with time to minimum strain value for each patient and
the selected strain segment
}
\description{
Find the time corresponding to the minimum strain value for a segment
}
\examples{
timeMinStrain(strain_data, "mas")
}
