% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_to_peak_strain.R
\name{calc_time_to_strain_peak}
\alias{calc_time_to_strain_peak}
\title{Calculate time to peak strain}
\usage{
calc_time_to_strain_peak(data, strain, time, thresh = 0, incr = 0,
  type = c("circumferential", "radial"), peak.criteria = c("units",
  "percent"), ...)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{strain}{Name of strain column, wrapped in "".}

\item{time}{Name of time column, wrapped in "".}

\item{thresh}{Optional, threshold that the strain peak has to pass.}

\item{incr}{Optional, difference between the sought after peak and the following peak of opposite direction.}

\item{type}{Type of strain, circumferential or radial.}

\item{peak.criteria}{Which criteria to use for peak.}
}
\description{
Calculate time to peak strain
}
