% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStrainPeaks.R
\name{findStrainPeaks}
\alias{findStrainPeaks}
\title{Finds peaks and valleys in the strain dataframe}
\usage{
findStrainPeaks(straindata, position.firstStraincolumn)
}
\arguments{
\item{straindata}{dataframe in longformat with separate column for each strain segment}

\item{position.firstStraincolumn}{the index position of the first strain column}
}
\value{
A dataframe with 2 new columns added for each strain segment indicating
whether the corresponding strain value is a peak/valley or not
}
\description{
Finds peaks and valleys in the strain dataframe
}
\examples{
findStrainPeaks(straindata, position.firstStrainColumn = 3)
}
