% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePeakStrain.R
\name{timePeakStrain}
\alias{timePeakStrain}
\title{Find the time corresponding to the peak strain value for a segment}
\usage{
timePeakStrain(data, id.column, time.column, RR.as.perc = TRUE,
  radial.strain = FALSE)
}
\arguments{
\item{data}{Dataframe with each strain segment in a separate column, the name
of the strain columns must include "strain".}

\item{id.column}{Name of the id column in quotes.}

\item{time.column}{Name of the time column in quotes.}

\item{RR.as.perc}{Logical parameter, if TRUE time will be presented as percent
of heart cycle.}

\item{radial.strain}{Logical parameter, set to TRUE if using radial strain.}
}
\value{
A dataframe with time to peak strain value for each patient and all
segments.
}
\description{
Find the time corresponding to the peak strain value for a segment
}
\examples{
timeMinStrain(strain_data, "id", "time", RR.as.perc = TRUE)
}
